35강. 애플리케이션 대신 DB로 기능 구현하기

+++ src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -2,4 +2,4 @@ package com.group.libraryapp.dto.book.response
 
 import com.group.libraryapp.type.BookType
 
-data class BookStatResponse(val type: BookType, val count: Int)
+data class BookStatResponse(val type: BookType, val count: Long)
